Boot: Move WordPress extraction logic to a common unzipWordPress() utility
TODO: Rewrite it in PHP using Zip Streams from the PHP Blueprints
library, find nested WordPress whether it's in a doubly compressed zip
archive or a "build" directory. Give the developer some control over it,
too, in case there are multiple zipped WordPresses. Or don't and
document that.
